requirements.txt
streamlit>=1.28.0
onnxruntime>=1.20.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
azure-storage-blob>=12.18.0
streamlit-drawable-canvas>=0.2.0
plotly>=5.17.0
